Add files via upload
Learning pen-testing. 
Second task was to setting up the Metasploitable VM on a host Kali Linux. Seeing my peers having trouble performing the task and I was the first one to complete the task after some hiccups and doing some searches on how to fix the issues, I decided to create an instructional document hoping to help them and others as well. 

Feel free to circulate but requesting not to change any information on the document. If the document has helped you, I would appreciate if you inform me about it or any update/mistakes that you may find in the document. This will give me a moral boost to continue doing things like this for the future.