Kaligpt offline integration
This file is an offline Kaligpt integration, where it downloads a model on your local machine and you get to use it without thinking of privacy issues.
Please do read the Read me file for intro and how to install the package on your linux machine.
With the help of chatgpt and lot of back forth prompts had to polish this project and it still requires more tweaking. Hopefully you will find it useful.